create PromotionalOfferViewTests
